requirements.txt
discord.py
openai==1.12.0
azure-identity
python-dotenv
pypresence
youtube_dl
pynacl